Use axios for API requests
